feat: Refactor log file management by introducing a new base class and platform-specific implementations for native and web platforms chore: Remove obsolete log file factory and related files, consolidating functionality into a more modular structure
