+++ packages/ispect/lib/src/common/utils/logs_file/base/base.dart
@@ -1,0 +1,4 @@
+/// Base abstract class for cross-platform log file operations.
+library;
+
+export 'base_logs_file.dart';
