+++ packages/ispect/lib/src/common/utils/logs_file/factory/factory.dart
@@ -1,0 +1,4 @@
+/// Factory classes for creating platform-appropriate log file handlers.
+library;
+
+export 'logs_file_factory.dart';
