+++ packages/ispect/lib/src/common/utils/logs_file/factory/logs_file_factory.dart
@@ -1,6 +1,6 @@
-import 'package:ispect/src/common/utils/logs_file/base_logs_file.dart';
-import 'package:ispect/src/common/utils/logs_file/logs_file_factory_native.dart'
-    if (dart.library.html) 'package:ispect/src/common/utils/logs_file/logs_file_factory_web.dart';
+import 'package:ispect/src/common/utils/logs_file/base/base_logs_file.dart';
+import 'package:ispect/src/common/utils/logs_file/factory/native_factory.dart'
+    if (dart.library.html) 'package:ispect/src/common/utils/logs_file/factory/web_factory.dart';
 
 /// Factory class for creating platform-appropriate log file handlers.
 ///
