+++ packages/ispect/lib/src/common/utils/logs_file/factory/web_factory.dart
@@ -1,0 +1,5 @@
+import 'package:ispect/src/common/utils/logs_file/base/base_logs_file.dart';
+import 'package:ispect/src/common/utils/logs_file/implementations/web_logs_file.dart';
+
+/// Web-specific factory implementation.
+BaseLogsFile createPlatformLogsFile() => WebLogsFile();
