+++ packages/ispect/lib/src/common/utils/logs_file/implementations/implementations.dart
@@ -1,0 +1,5 @@
+/// Platform-specific implementations of log file operations.
+library;
+
+export 'native_logs_file.dart';
+export 'web_logs_file.dart';
