+++ packages/ispect/lib/src/common/utils/logs_file/logs_file.dart
@@ -1,7 +1,9 @@
 // Cross-platform logs file utilities
-export 'base_logs_file.dart';
-export 'logs_file_factory.dart';
+export 'base/base.dart';
+export 'factory/factory.dart';
 
 // Export platform-specific implementations
-export 'native_logs_file.dart' if (dart.library.html) 'web_logs_file.dart';
-export 'web_logs_file.dart' if (dart.library.io) 'native_logs_file.dart';
+export 'implementations/native_logs_file.dart'
+    if (dart.library.js_interop) 'implementations/web_logs_file.dart';
+export 'implementations/web_logs_file.dart'
+    if (dart.library.io) 'implementations/native_logs_file.dart';
