--- packages/ispect/lib/src/common/utils/logs_file/logs_file_factory_native.dart
@@ -1,5 +1,0 @@
-import 'package:ispect/src/common/utils/logs_file/base_logs_file.dart';
-import 'package:ispect/src/common/utils/logs_file/native_logs_file.dart';
-
-/// Native-specific factory implementation.
-BaseLogsFile createPlatformLogsFile() => NativeLogsFile();
