--- packages/ispect/lib/src/common/utils/logs_file/logs_file_factory_web.dart
@@ -1,5 +1,0 @@
-import 'package:ispect/src/common/utils/logs_file/base_logs_file.dart';
-import 'package:ispect/src/common/utils/logs_file/web_logs_file.dart';
-
-/// Web-specific factory implementation.
-BaseLogsFile createPlatformLogsFile() => WebLogsFile();
